+++ SplunkCustomTracking/Sources/SplunkCustomTracking/Module/CustomTracking+Module.swift
@@ -28,8 +28,10 @@ extension CustomTrackingInternal: Module {
     public typealias EventMetadata = CustomTrackingMetadata
     public typealias EventData = CustomTrackingData
 
-    public func install(with _: (any ModuleConfiguration)?, remoteConfiguration _: (any SplunkCommon.RemoteModuleConfiguration)?) {
-        _ = Self.instance
+    public func install(with configuration: (any ModuleConfiguration)?, remoteConfiguration _: (any SplunkCommon.RemoteModuleConfiguration)?) {
+        if let configuration = configuration as? CustomTrackingConfiguration {
+            diagnosticEnricher.configure(includeBinaryImagesOnErrors: configuration.includeBinaryImagesOnErrors)
+        }
     }
 
     public func onPublish(data: @escaping (CustomTrackingMetadata, CustomTrackingData) -> Void) {